fix: CodeRabbitレビュー指摘4件を修正
- Go(gin/echo/beego): 全ハンドラーの limit に上限500チェックを追加（オーバーフロー対策）
- Kotlin: buildPager/DEFAULT_PAGE_COUNT を PagerUtils.kt に集約し重複を解消
- Kotlin: ExposedJwtHistoryRepository・ExposedStaffRepository の .limit() に minOf(...,500) を追加
- Ruby Rails: sort_type の比較を .to_s.downcase で大文字小文字非依存に修正
