+++ backends/go-echo/internal/handler/client_handler.go
@@ -57,6 +57,9 @@ func (h *ClientHandler) Index(c echo.Context) error {
 			limit = n
 		}
 	}
+	if limit > 500 {
+		limit = 500
+	}
 	page := 1
 	if v := c.QueryParam("page"); v != "" {
 		if n, err := strconv.Atoi(v); err == nil && n > 0 {
@@ -241,6 +244,9 @@ func (h *ClientHandler) JwtHistories(c echo.Context) error {
 		if n, err := strconv.Atoi(v); err == nil && n > 0 {
 			limit = n
 		}
+	}
+	if limit > 500 {
+		limit = 500
 	}
 	page := 1
 	if v := c.QueryParam("page"); v != "" {
