+++ backends/go-echo/internal/handler/staff_handler.go
@@ -34,6 +34,9 @@ func (h *StaffHandler) Index(c echo.Context) error {
 			limit = n
 		}
 	}
+	if limit > 500 {
+		limit = 500
+	}
 	page := 1
 	if v := c.QueryParam("page"); v != "" {
 		if n, err := strconv.Atoi(v); err == nil && n > 0 {
