+++ backends/go-gin/internal/handler/staff_handler.go
@@ -42,6 +42,9 @@ func (h *StaffHandler) Index(c *gin.Context) {
 			limit = n
 		}
 	}
+	if limit > 500 {
+		limit = 500
+	}
 	page := 1
 	if v := c.Query("page"); v != "" {
 		if n, err := strconv.Atoi(v); err == nil && n > 0 {
